Remove compiler warnings

+++ src/ir/ir.parse.hpp
@@ -9,7 +9,7 @@ char* irParseString(byte* payload);
 
 IRSetting* irParsePower(bool power);
 IRSetting* irParseMode(const char* value);
-IRSetting* irParseTemp(int value);
+IRSetting* irParseTemp(uint8_t value);
 IRSetting* irParseFanSpeed(const char* value);
 IRSetting* irParseFanDirVert(const char* value);
 IRSetting* irParseFanDirHorz(const char* value);
